removed multiclosure_inconsistent and moved to multiclosure
